services/users/requirements/production.txt
## This file is empty
Flask~=1.1.2
click~=7.1.2
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate